Recover when syslog-ng is restarted
When syslog-ng restart, the socket on disk is removed and a new one is
created.  On the riemann-tool side, this result in an EPIPE exception
being raised.

When this happen, initiate a new connection and retry the operation.

If the socket is not available at startup, we should not terminate
execution, so rework the socket initialization logic so that it happens
"when needed" rather that "on startup".  Any ECONNREFUSED or EACCES
exception will be output to stderr but execution will continue.